fix(runtime): share running record validation
