+++ src/runtime/src/local_execution/session.rs
@@ -10,8 +10,7 @@ use a3s_box_core::{
 };
 use async_trait::async_trait;
 
-use super::support::{managed_state, require_generation};
-use super::{LocalExecutionManager, ManagedExecutionState};
+use super::LocalExecutionManager;
 use crate::{
     BoxRecord, ExecClient, PtyClient, StreamingExec, StreamingExecInput, StreamingPty,
     StreamingPtyInput,
@@ -92,39 +91,6 @@ impl ExecutionSessionManager for LocalExecutionManager {
 }
 
 impl LocalExecutionManager {
-    pub(super) async fn require_running_record(
-        &self,
-        execution_id: &ExecutionId,
-        generation: ExecutionGeneration,
-    ) -> ExecutionManagerResult<BoxRecord> {
-        let record = self
-            .get(execution_id)
-            .await?
-            .ok_or_else(|| ExecutionManagerError::NotFound(execution_id.clone()))?;
-        require_generation(&record, execution_id, generation)?;
-        if managed_state(&record)? != ManagedExecutionState::Running {
-            return Err(ExecutionManagerError::Conflict {
-                execution_id: execution_id.clone(),
-                message: "execution is not running".to_string(),
-            });
-        }
-        if record.exec_socket_path.as_os_str().is_empty() {
-            return Err(ExecutionManagerError::Internal(format!(
-                "execution {execution_id} has no exec endpoint"
-            )));
-        }
-        #[cfg(target_os = "linux")]
-        {
-            let pid = record
-                .pid
-                .ok_or_else(|| ExecutionManagerError::NotFound(execution_id.clone()))?;
-            if !crate::process::is_process_alive_with_identity(pid, record.pid_start_time) {
-                return Err(ExecutionManagerError::NotFound(execution_id.clone()));
-            }
-        }
-        Ok(record)
-    }
-
     async fn bind_exec(
         &self,
         execution_id: &ExecutionId,
